Switched from float32 to float64

+++ pgf_kernel_experiments/experiments/earth_temperature/tune_multiple_gps.py
@@ -177,11 +177,11 @@ test_output = dataset['tempanomaly'][t, :, :].flatten()[[test_ids]].squeeze()
 
 # %% Convert training and test data to PyTorch format
 
-train_x = torch.as_tensor(train_pos, dtype=torch.float32)
-train_y = torch.as_tensor(train_output, dtype=torch.float32)
+train_x = torch.as_tensor(train_pos, dtype=torch.float64)
+train_y = torch.as_tensor(train_output, dtype=torch.float64)
 
-test_x = torch.as_tensor(test_pos, dtype=torch.float32)
-test_y = torch.as_tensor(test_output, dtype=torch.float32)
+test_x = torch.as_tensor(test_pos, dtype=torch.float64)
+test_y = torch.as_tensor(test_output, dtype=torch.float64)
 
 # %% Set up ExactMultiGPRunner
